Raise delete and archive to AbstractModuleCollection level
