+++ src/citrine/resources/design_space.py
@@ -1,9 +1,8 @@
 """Resources that represent collections of design spaces."""
 from uuid import UUID
-from typing import TypeVar, Union
+from typing import TypeVar
 
 from citrine._session import Session
-from citrine.exceptions import CitrineException
 from citrine.resources.module import AbstractModuleCollection
 from citrine.informatics.design_spaces import DesignSpace, EnumeratedDesignSpace
 
@@ -63,21 +62,6 @@ class DesignSpaceCollection(AbstractModuleCollection[DesignSpace]):
         """Update an existing design space by uid."""
         self.validate_write_request(model)
         return AbstractModuleCollection.update(self, model)
-
-    def delete(self, uid: Union[UUID, str]):
-        """Modules cannot be deleted at this time."""
-        msg = "Design spaces cannot be deleted at this time. Use 'archive' instead."
-        raise NotImplementedError(msg)
-
-    def archive(self, module_id: Union[UUID, str]) -> DesignSpace:
-        """Archiving a design space removes it from view, but is not a hard delete."""
-        try:
-            module = self.get(module_id)
-        except CitrineException:
-            msg = f"Design space with id {module_id} was not found, and hence cannot be archived."
-            raise RuntimeError(msg)
-        module.archived = True
-        return self.update(module)
 
     def create_default(self, predictor_id: UUID) -> DesignSpace:
         """[ALPHA] Create a default design space for a predictor.
