+++ src/citrine/resources/module.py
@@ -1,9 +1,10 @@
 """Resources that represent collections of design spaces."""
 from uuid import UUID
-from typing import TypeVar
+from typing import TypeVar, Union
 
 from citrine._rest.collection import Collection
 from citrine._session import Session
+from citrine.exceptions import CitrineException
 from citrine.informatics.modules import Module
 
 
@@ -37,3 +38,20 @@ ModuleType = TypeVar('ModuleType', bound='Module')
 
 class AbstractModuleCollection(Collection[ModuleType]):
     """Abstract class for representing collection of modules."""
+
+    _module_type = None
+
+    def delete(self, uid: Union[UUID, str]):
+        """Modules cannot be deleted at this time."""
+        msg = "Modules cannot be deleted at this time. Use 'archive' instead."
+        raise NotImplementedError(msg)
+
+    def archive(self, module_id: Union[UUID, str]) -> ModuleType:
+        """Archiving amodule removes it from view, but is not a hard delete."""
+        try:
+            module = self.get(module_id)
+        except CitrineException:
+            raise RuntimeError(f"{self._module_type} with id {module_id} was not found, "
+                               f"and hence cannot be archived.")
+        module.archived = True
+        return self.update(module)
