+++ src/citrine/resources/processor.py
@@ -1,9 +1,8 @@
 """Resources that represent collections of processors."""
 from uuid import UUID
-from typing import TypeVar, Union
+from typing import TypeVar
 
 from citrine._session import Session
-from citrine.exceptions import CitrineException
 from citrine.resources.module import AbstractModuleCollection
 from citrine.informatics.processors import Processor
 
@@ -34,18 +33,3 @@ class ProcessorCollection(AbstractModuleCollection[Processor]):
         processor = Processor.build(data)
         processor.session = self.session
         return processor
-
-    def delete(self, uid: Union[UUID, str]):
-        """Modules cannot be deleted at this time."""
-        msg = "Processors cannot be deleted at this time. Use 'archive' instead."
-        raise NotImplementedError(msg)
-
-    def archive(self, module_id: Union[UUID, str]) -> Processor:
-        """Archiving a processor removes it from view, but is not a hard delete."""
-        try:
-            module = self.get(module_id)
-        except CitrineException:
-            msg = f"Processor with id {module_id} was not found, and hence cannot be archived."
-            raise RuntimeError(msg)
-        module.archived = True
-        return self.update(module)
